refactor: move fpage methods to fpage.c
